move domain service configuration into domain package
